feat: moved functions from orgview to relevant component
